Added my name as author vs zf297a Added a revision comment Made sure the env var defaults to /apps/CRON/AMD/lib
Changed the usage conditional to use the numeric expression and a
separate AND'ed contion that checks if param 1 is a question mark.

in the abort function, check the env var debug and if it equals Y turn
on the displaying of variable substitutions for the function by using:
set -x.  Even if set -x is turned on globally, it does not automatically
get turned on for funtions.  set -x must be invoked for functions if
their "environment" variables are to be displayed.

Added function loadFlatFiles - the FDW, finance data warehouse, data
supplied by Dana Ernst, the load of the Warner Robins data, and the
benchstock data.  All these are optional and can be loaded via
SQL*Loader from a PC.

For function execSqlplus, added a DEBUGOPT env var that gets set to the
empty string, but if debug = Y then it becomes -d and gets passed to
execSqlplus.ksh as the first argument - this turns on debugging for that
script.

Use the == when checking string equality vs the deprecated =